requirements.txt
fastapi
uvicorn[standard]
streamlit
google-cloud-bigquery
google-cloud-storage
google-cloud-aiplatform
python-jose[cryptography]
passlib[bcrypt]
pandas
requests
tenacity
streamlit-authenticator
reportlab
pytest-playwright
SQLAlchemy
google-generativeai
beautifulsoup4
flask